+++ app/scripts/components/index.ts
@@ -1,3 +1,4 @@
 import "./hierarchy-path-canvas.component";
 import "./level.component";
 import "./bundlelist-sidebar.component";
+import "./emuhierarchy.component";