+++ app/NEWS.md
@@ -1,5 +1,17 @@
 # What's New
 
+## Version 1.2.6.9000
+
+### bug fixes
+
+
+### new features / performance tweaks / improvements
+
+- fixed help if comMode ist GITLAB
+- fixed requestAnimationFrame binding of this
+- D3 hierarchy SVG now a component
+
+
 ## Version 1.2.5
 
 ### bug fixes